+++ qbit/core/src/main/java/io/advantageous/qbit/queue/impl/sender/AbstractBasicSendQueue.java
@@ -1,4 +1,4 @@
-package io.advantageous.qbit.queue.impl;
+package io.advantageous.qbit.queue.impl.sender;
 
 import io.advantageous.qbit.queue.Queue;
 import io.advantageous.qbit.queue.SendQueue;
@@ -7,8 +7,9 @@ import org.slf4j.Logger;
 import java.util.Collection;
 import java.util.concurrent.BlockingQueue;
 
-import static io.advantageous.qbit.queue.impl.SenderHelperMethods.objectArray;
-import static io.advantageous.qbit.queue.impl.SenderHelperMethods.objectArrayFromIterable;
+import static io.advantageous.qbit.queue.impl.sender.SenderHelperMethods.fastObjectArraySlice;
+import static io.advantageous.qbit.queue.impl.sender.SenderHelperMethods.objectArray;
+import static io.advantageous.qbit.queue.impl.sender.SenderHelperMethods.objectArrayFromIterable;
 
 public abstract class AbstractBasicSendQueue <T> implements SendQueue<T> {
 
@@ -34,6 +35,8 @@ public abstract class AbstractBasicSendQueue <T> implements SendQueue<T> {
                 
     }
 
+    protected abstract boolean flushIfOverBatch();
+    protected abstract boolean sendArray(Object[] items);
 
     public boolean shouldBatch() {
         return true;
@@ -41,7 +44,7 @@ public abstract class AbstractBasicSendQueue <T> implements SendQueue<T> {
 
 
     @Override
-    public boolean send(T item) {
+    public final boolean send(T item) {
         checkStarted();
         boolean ableToSend = flushIfOverBatch();
         queueLocal[index] = item;
@@ -49,11 +52,8 @@ public abstract class AbstractBasicSendQueue <T> implements SendQueue<T> {
         return ableToSend;
     }
 
-    protected abstract boolean flushIfOverBatch();
-    protected abstract boolean sendArray(Object[] items);
-    protected abstract boolean sendLocalQueue();
 
-    private void checkStarted() {
+    private final void checkStarted() {
 
         if (checkEveryStarted % 100 == 0) {
             if (!owner.started()) {
@@ -65,7 +65,7 @@ public abstract class AbstractBasicSendQueue <T> implements SendQueue<T> {
     }
 
     @Override
-    public void sendAndFlush(T item) {
+    public final void sendAndFlush(T item) {
         checkStarted();
         send(item);
         flushSends();
@@ -81,7 +81,7 @@ public abstract class AbstractBasicSendQueue <T> implements SendQueue<T> {
 
 
     @Override
-    public void sendBatch(Iterable<T> items) {
+    public final void sendBatch(Iterable<T> items) {
         checkStarted();
         flushSends();
         final Object[] array = objectArrayFromIterable(items);
@@ -89,7 +89,7 @@ public abstract class AbstractBasicSendQueue <T> implements SendQueue<T> {
     }
 
     @Override
-    public void sendBatch(Collection<T> items) {
+    public final void sendBatch(Collection<T> items) {
         checkStarted();
         flushSends();
         final Object[] array = objectArray(items);
@@ -99,26 +99,33 @@ public abstract class AbstractBasicSendQueue <T> implements SendQueue<T> {
 
 
     @Override
-    public int hashCode() {
+    public final int hashCode() {
         return queue.hashCode();
     }
 
     @Override
-    public String name() {
+    public final String name() {
         return name;
     }
 
 
     @Override
-    public int size() {
+    public final int size() {
         return queue.size();
     }
 
     @Override
-    public void flushSends() {
+    public final void flushSends() {
         if (index > 0) {
             sendLocalQueue();
         }
     }
 
+    protected final boolean sendLocalQueue() {
+
+        final Object[] copy = fastObjectArraySlice(queueLocal, 0, index);
+        boolean ableToSend = sendArray(copy);
+        index = 0;
+        return ableToSend;
+    }
 }
